Now reacts to a click event on an image
I have spend the whole of yesterday AND today researching how to do this
and trying out several dozens variations of the same code.

It worked after I removed the dang DOMContentLoaded event check.  This
is no longer necessary anyway, because of the 'run_at' property in the
manifest.json under 'content_scripts'.